requirements.txt
streamlit
arxiv
langchain-groq
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-core
chromadb
python-dotenv
